New page SignUp implemented+added to Router#24
