+++ src/shared/api/constants/countries.json
@@ -150,7 +150,6 @@
     {"name": "Paraguay", "cca2": "PY", "flag": "🇵🇾", "code": "595"},
     {"name": "Qatar", "cca2": "QA", "flag": "🇶🇦", "code": "974"},
     {"name": "Romania", "cca2": "RO", "flag": "🇷🇴", "code": "40"},
-    {"name": "Russia", "cca2": "RU", "flag": "🇷🇺", "code": "7"},
     {"name": "Rwanda", "cca2": "RW", "flag": "🇷🇼", "code": "250"},
     {"name": "Saudi Arabia", "cca2": "SA", "flag": "🇸🇦", "code": "966"},
     {"name": "Sudan", "cca2": "SD", "flag": "🇸🇩", "code": "249"},
